requirements.txt
appdirs==1.4.3
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.4
django-allauth==0.41.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
oauthlib==3.1.0
psycopg2==2.7.5
psycopg2-binary==2.8.4
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.12.0
sqlparse==0.3.1
urllib3==1.25.8
virtualenv==20.0.8
whitenoise==5.0.1
zipp==3.1.0
